modification in several files: add getter and setters, methods...

+++ projects/back/src/main/java/com/exerc/mybackend/repositories/HousingLocationRepository.java
@@ -3,5 +3,5 @@ package com.exerc.mybackend.repositories;
 import com.exerc.mybackend.entities.HousingLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HousingLocationRepository extends JpaRepository<HousingLocation, Integer> {
+public interface HousingLocationRepository extends JpaRepository<HousingLocation, Long> {
 }
